添加控制层

+++ soukon-auto-work/src/main/java/com/soukon/domain/VOS/DataCellVO.java
@@ -10,6 +10,6 @@ import java.util.List;
 public class DataCellVO {
     private long id;
     private String name;
-    private List<Long> value;
+    private List<Double> value;
 
 }
